Update user model
